+++ plugins/mp3/mp3.c
@@ -116,7 +116,7 @@ cmp3_set_extra_properties (DB_playItem_t *it, mp3info_t *mp3info, int fake) {
     char s[100];
     int64_t size = mp3info->fsize;
     if (size >= 0) {
-        snprintf (s, sizeof (s), "%lld", size);
+        snprintf (s, sizeof (s), "%lld", (long long)size);
         deadbeef->pl_replace_meta (it, ":FILE_SIZE", s);
     }
     else {
